+++ src/utils/getSummaryDetails.ts
@@ -3,6 +3,7 @@ import { getTotalStatus } from ".";
 
 export const getSummaryDetails = (suite: Suite): string => {
   const totalStatus = getTotalStatus(suite.suites);
+  const totalFailed = totalStatus.failed + totalStatus.timedOut;
 
   const headerText = [`<li>Total tests: ${suite.allTests().length}</li>`];
 
@@ -10,8 +11,8 @@ export const getSummaryDetails = (suite: Suite): string => {
     headerText.push(`<li>Passed: ${totalStatus.passed}</li>`);
   }
 
-  if (totalStatus.failed > 0) {
-    headerText.push(`<li>Failed: ${totalStatus.failed}</li>`);
+  if (totalFailed > 0) {
+    headerText.push(`<li>Failed: ${totalFailed}</li>`);
   }
 
   if (totalStatus.skipped > 0) {
@@ -19,7 +20,7 @@ export const getSummaryDetails = (suite: Suite): string => {
   }
 
   if (totalStatus.timedOut > 0) {
-    headerText.push(`<li>Timed Out: ${totalStatus.timedOut}</li>`);
+    headerText.push(`<li>Timed out: ${totalStatus.timedOut}</li>`);
   }
 
   return `<ul>${headerText.join("")}</ul>`;